Add .NET 8 version and remove .NET 3 and 6 versions
