--- backend/csharpdotnet-3/dotnetcore3.sln
@@ -1,36 +1,0 @@
-
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 25.0.1703.3
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dotnetcore3", "dotnetcore3\dotnetcore3.csproj", "{2570ACBB-701B-4F45-AE96-881C45866E89}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dotnetcore3_tests", "dotnetcore3_tests\dotnetcore3_tests.csproj", "{D3277DC4-DBDF-4868-B1A8-DA3A36DADCA0}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{2570ACBB-701B-4F45-AE96-881C45866E89}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2570ACBB-701B-4F45-AE96-881C45866E89}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2570ACBB-701B-4F45-AE96-881C45866E89}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2570ACBB-701B-4F45-AE96-881C45866E89}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D3277DC4-DBDF-4868-B1A8-DA3A36DADCA0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D3277DC4-DBDF-4868-B1A8-DA3A36DADCA0}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D3277DC4-DBDF-4868-B1A8-DA3A36DADCA0}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D3277DC4-DBDF-4868-B1A8-DA3A36DADCA0}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FAE624B4-786C-4799-825F-F0C6721A7526}
-	EndGlobalSection
-	GlobalSection(MonoDevelopProperties) = preSolution
-		Policies = $0
-		$0.DotNetNamingPolicy = $1
-		$1.DirectoryNamespaceAssociation = PrefixedHierarchical
-	EndGlobalSection
-EndGlobal
