--- backend/csharpdotnet-3/dotnetcore3/Interfaces/ICartService.cs
@@ -1,12 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using dotnetcore3.Models;
-
-namespace dotnetcore3.Interfaces
-{
-    public interface ICartService
-    {
-        public List<Product> GetProducts();
-    }
-}
-
