--- backend/csharpdotnet-3/dotnetcore3/Models/ItemsResponse.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-namespace dotnetcore3.Models
-{
-    public class ItemsResponse
-    {
-        public List<Product> ProductList { get; set; }
-
-        public ItemsResponse(List<Product> productList)
-        {
-            ProductList = productList;
-        }
-    }
-}
-
