--- backend/csharpdotnet-3/dotnetcore3/Models/Product.cs
@@ -1,17 +1,0 @@
-﻿using System;
-namespace dotnetcore3.Models
-{
-    public class Product
-    {
-        public string ItemName { get; set; }
-
-        public int Price { get; set; }
-
-        public Product(string item_name, int price)
-        {
-            ItemName = item_name;
-            Price = price;
-        }
-    }
-}
-
