--- backend/csharpdotnet-3/dotnetcore3/appsettings.Development.json
@@ -1,10 +1,0 @@
-﻿{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  }
-}
-
