--- backend/csharpdotnet-3/dotnetcore3/dotnetcore3.csproj
@@ -1,19 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <AssemblyName>dotnetcore3</AssemblyName>
-  </PropertyGroup>
-
-
-  <ItemGroup>
-    <None Remove="Models\" />
-    <None Remove="Interfaces\" />
-    <None Remove="Services\" />
-  </ItemGroup>
-  <ItemGroup>
-    <Folder Include="Models\" />
-    <Folder Include="Interfaces\" />
-    <Folder Include="Services\" />
-  </ItemGroup>
-</Project>
