--- backend/csharpdotnet-6/dotnet6.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 25.0.1703.2
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dotnet6_tests", "dotnet6_tests\dotnet6_tests.csproj", "{8A22A9BB-54D3-4669-BC4B-34C844CC43F1}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dotnet6", "dotnet6\dotnet6.csproj", "{427E4E38-5F29-437E-9911-812447690B1D}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{8A22A9BB-54D3-4669-BC4B-34C844CC43F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8A22A9BB-54D3-4669-BC4B-34C844CC43F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8A22A9BB-54D3-4669-BC4B-34C844CC43F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8A22A9BB-54D3-4669-BC4B-34C844CC43F1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{427E4E38-5F29-437E-9911-812447690B1D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{427E4E38-5F29-437E-9911-812447690B1D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{427E4E38-5F29-437E-9911-812447690B1D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{427E4E38-5F29-437E-9911-812447690B1D}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CEFBAEFE-4702-475E-94F7-4836DAFB0052}
-	EndGlobalSection
-EndGlobal
