--- backend/csharpdotnet-6/dotnet6/Interfaces/ICartService.cs
@@ -1,11 +1,0 @@
-﻿using System;
-using dotnet6.Models;
-
-namespace dotnet6.Interfaces
-{
-    public interface ICartService
-    {
-        public List<Product> GetProducts();
-    }
-}
-
