--- backend/csharpdotnet-6/dotnet6/Models/ItemsResponse.cs
@@ -1,14 +1,0 @@
-﻿using System;
-namespace dotnet6.Models
-{
-    public class ItemsResponse
-    {
-        public List<Product> ProductList { get; set; }
-
-        public ItemsResponse(List<Product> productList)
-        {
-            ProductList = productList;
-        }
-    }
-}
-
