--- backend/csharpdotnet-6/dotnet6/Models/Product.cs
@@ -1,17 +1,0 @@
-﻿using System;
-namespace dotnet6.Models
-{
-    public class Product
-    {
-        public string ItemName { get; set; }
-
-        public int Price { get; set; }
-
-        public Product(string item_name, int price)
-        {
-            ItemName = item_name;
-            Price = price;
-        }
-    }
-}
-
