--- backend/csharpdotnet-6/dotnet6_tests/Usings.cs
@@ -1,1 +1,0 @@
-﻿global using Xunit;
