+++ backend/csharpdotnet-8/dotnet8.sln
@@ -1,0 +1,27 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.31903.59
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dotnet8", "dotnet8\dotnet8.csproj", "{A1B2C3D4-E5F6-4A5B-8C7D-9E0F1A2B3C4D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "dotnet8_tests", "dotnet8_tests\dotnet8_tests.csproj", "{D4C3B2A1-F6E5-B5A4-D7C8-E3F2A1B0C9D8}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A1B2C3D4-E5F6-4A5B-8C7D-9E0F1A2B3C4D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A1B2C3D4-E5F6-4A5B-8C7D-9E0F1A2B3C4D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A1B2C3D4-E5F6-4A5B-8C7D-9E0F1A2B3C4D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A1B2C3D4-E5F6-4A5B-8C7D-9E0F1A2B3C4D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D4C3B2A1-F6E5-B5A4-D7C8-E3F2A1B0C9D8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D4C3B2A1-F6E5-B5A4-D7C8-E3F2A1B0C9D8}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D4C3B2A1-F6E5-B5A4-D7C8-E3F2A1B0C9D8}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D4C3B2A1-F6E5-B5A4-D7C8-E3F2A1B0C9D8}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
