+++ backend/csharpdotnet-8/dotnet8/Interfaces/ICartService.cs
@@ -1,0 +1,11 @@
+using dotnet8.Models;
+
+namespace dotnet8.Interfaces;
+
+public interface ICartService
+{
+    List<Product> GetProducts();
+    
+    // New method to calculate total price
+    double CalculateTotalPrice(string[] items);
+}
