+++ backend/csharpdotnet-8/dotnet8/Models/ItemsResponse.cs
@@ -1,0 +1,11 @@
+namespace dotnet8.Models;
+
+public class ItemsResponse
+{
+    public List<Product> ProductList { get; set; } = new();
+
+    public ItemsResponse(List<Product> productList)
+    {
+        ProductList = productList;
+    }
+}
