+++ backend/csharpdotnet-8/dotnet8/Models/Product.cs
@@ -1,0 +1,14 @@
+namespace dotnet8.Models;
+
+public class Product
+{
+    public string ItemName { get; set; } = string.Empty;
+
+    public int Price { get; set; }
+
+    public Product(string itemName, int price)
+    {
+        ItemName = itemName;
+        Price = price;
+    }
+}
