+++ backend/csharpdotnet-8/dotnet8/Program.cs
@@ -1,8 +1,9 @@
-﻿using dotnet6.Interfaces;
-using dotnet6.Services;
+using dotnet8.Interfaces;
+using dotnet8.Services;
 
 var builder = WebApplication.CreateBuilder(args);
 
+// Add services to the container
 builder.Services.AddControllers()
                 .AddJsonOptions(options =>
                 {
@@ -11,16 +12,27 @@ builder.Services.AddControllers()
                     options.JsonSerializerOptions.IncludeFields = true;
                 });
 
+// Add Swagger/OpenAPI support
+builder.Services.AddEndpointsApiExplorer();
+builder.Services.AddSwaggerGen();
+
 // Add services
 builder.Services.AddSingleton<ICartService, CartService>();
 
+// Configure the HTTP request pipeline
 builder.WebHost.UseUrls("http://localhost:5050");
 
 var app = builder.Build();
 
-app.UseAuthorization();
+// Configure middleware
+if (app.Environment.IsDevelopment())
+{
+    app.UseSwagger();
+    app.UseSwaggerUI();
+}
 
+app.UseHttpsRedirection();
+app.UseAuthorization();
 app.MapControllers();
 
-app.Run();
-
+app.Run();
