+++ backend/csharpdotnet-8/dotnet8/appsettings.Development.json
@@ -1,9 +1,8 @@
-﻿{
+{
   "Logging": {
     "LogLevel": {
       "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
+      "Microsoft.AspNetCore": "Information"
     }
   }
-}
-
+}
