+++ backend/csharpdotnet-8/dotnet8/appsettings.json
@@ -1,11 +1,9 @@
-﻿{
+{
   "Logging": {
     "LogLevel": {
       "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
+      "Microsoft.AspNetCore": "Warning"
     }
   },
   "AllowedHosts": "*"
-}
-
+}
