+++ backend/csharpdotnet-8/dotnet8_tests/Usings.cs
@@ -1,0 +1,1 @@
+global using Xunit;
